+++ src/index.ts
@@ -1,5 +1,13 @@
+// Converters related utils
+export { default as convertTextToSlug } from './utils/converters/convertTextToSlug.ts'
+
 // Data related utils
 export { default as extractBase64Data } from './utils/data/extractBase64Data.ts'
+
+// Generators related utils
+export { default as generateCode } from './utils/generators/generateCode.ts'
+export { default as generateId } from './utils/generators/generateId.ts'
+export { default as generateNumberBetween } from './utils/generators/generateNumberBetween.ts'
 
 // Time related utils
 export { default as isTimestampExpired } from './utils/time/isTimestampExpired.ts'